+++ packages/blip/app/core/api-utils.js
@@ -37,11 +37,8 @@
  * @typedef { import("./lib/partial-data-load").DateRange } DateRange
  */
 
-import _ from "lodash";
 import bows from "bows";
 import moment from "moment-timezone";
-
-import { DAILY_TYPES } from "tideline";
 
 import PartialDataLoad from "./lib/partial-data-load";
 
@@ -57,31 +54,22 @@ class ApiUtils {
     /** @type {Console} */
     this.log = bows("ApiUtils");
 
-    /** @type {[moment.Moment, moment.Moment]} */
-    this.dataRange = null;
     /** @type {PartialDataLoad} */
     this.partialDataLoad = null;
+  }
 
-    this.logWarnNoAPIv1 = _.once((err) => this.log.warn("Data API v1 not available", err.message));
-    this.haveAPIv1 = true;
+  get dateRange() {
+    return this.partialDataLoad.range;
   }
 
   /**
    * Refresh the data -> discard the previous loading information
    * @returns {Promise<PatientData>} The patient data
    */
-  async refresh() {
-    this.dataRange = null;
+  refresh() {
     this.partialDataLoad = null;
 
-    try {
-      return await this.refreshV1();
-    } catch (err) {
-      this.logWarnNoAPIv1(err);
-      this.haveAPIv1 = false;
-    }
-
-    return this.refreshV0();
+    return this.refreshV1();
   }
 
   /**
@@ -94,19 +82,22 @@ class ApiUtils {
       this.log.info("Range is empty - no data available");
       return [];
     }
-    this.log.info("Available data range:", range[0], range[1]);
 
-    // Assume browser locale, will adjust after
-    const timezone = new Intl.DateTimeFormat().resolvedOptions().timeZone;
-    this.dataRange = [moment.tz(range[0], timezone), moment.tz(range[1], timezone)];
+    const start = moment.utc(range[0]).startOf("day");
+    const end = moment.utc(range[1]).startOf("day");
+
+    this.log.info("Available data range:",
+      range[0], range[1],
+      "updated to", start.toISOString(), end.toISOString()
+    );
 
     // Get the initial range of data to load:
     // 3 weeks (for basics view) -> start/end of week
-    // substract one day to be sure to have all the data we need
+    // subtract one day to be sure to have all the data we need
     // since the timezone is generally not UTC
     const initialLoadingDates = [
-      moment.utc(this.dataRange[1]).startOf("week").subtract(2, "weeks").subtract(1, "day"),
-      moment.utc(this.dataRange[1]).endOf("day"),
+      end.clone().startOf("week").subtract(2, "weeks").subtract(1, "day"),
+      end,
     ];
 
     /** @type {GetPatientDataOptions} */
@@ -115,8 +106,6 @@ class ApiUtils {
       withPumpSettings: true,
     };
 
-    this.log.info("Initial loading options:", loadingOptions);
-
     // Get the data from the API
     const [patientData, messagesNotes] = await Promise.all([
       this.api.getPatientData(this.patient, loadingOptions),
@@ -124,78 +113,11 @@ class ApiUtils {
     ]);
 
     this.partialDataLoad = new PartialDataLoad(
-      { start: this.dataRange[0].valueOf(), end: this.dataRange[1].valueOf() },
-      { start: initialLoadingDates[0].valueOf(), end: initialLoadingDates[1].valueOf() }
+      { start, end },
+      { start: initialLoadingDates[0], end: initialLoadingDates[1] }
     );
 
     return patientData.concat(messagesNotes);
-  }
-
-  /**
-   * Try do do the same thing than the APIv1 but using the v0 API
-   * @returns {Promise<PatientData>} The patient data
-   * @private
-   */
-  async refreshV0() {
-    // To get the data range we will start to fetch the upload data
-    // Since everything needs it
-    // The we will fetch the latest data of each types
-
-    /** @type {[PatientData, PatientData]} */
-    const [uploads, latest] = await Promise.all([
-      await this.api.getPatientDataV0(this.patient, { types: ["upload"] }),
-      await this.api.getPatientDataV0(this.patient, { latest: true }),
-    ]);
-
-    // Combine
-    let combinedData = uploads.concat(latest);
-    if (combinedData.length > 0) {
-      // Sort (using a field that's will be stripped after by TidelineData)
-      // Don't use epoch here because TidelineData use it to know if it
-      // has already process it.
-      combinedData.forEach((d) => {
-        d.timeProcessing = Date.parse(d.time);
-      });
-      combinedData.sort((a, b) => a.timeProcessing - b.timeProcessing);
-
-      this.log.debug("combinedData", combinedData);
-
-      const timezone = new Intl.DateTimeFormat().resolvedOptions().timeZone;
-      this.dataRange = [
-        moment.tz(combinedData[0].timeProcessing, timezone),
-        moment.tz(_.findLast(combinedData, (d) => DAILY_TYPES.includes(d.type)).timeProcessing, timezone),
-      ];
-
-      this.log.info("Available data range:", this.dataRange[0].toISOString(), this.dataRange[1].toISOString());
-
-      const initialLoadingDates = [
-        moment.utc(this.dataRange[1]).startOf("week").subtract(2, "weeks").subtract(1, "day"),
-        moment.utc(this.dataRange[1]).endOf("day"),
-      ];
-
-      /** @type {GetPatientDataOptionsV0} */
-      const loadingOptions = {
-        startDate: initialLoadingDates[0].toISOString(),
-      };
-
-      this.log.info("Initial loading options:", loadingOptions);
-
-      // Get the initial data from the API
-      /** @type {[PatientData, PatientData]} */
-      const [patientData, messagesNotes] = await Promise.all([
-        this.api.getPatientDataV0(this.patient, loadingOptions),
-        this.api.getMessages(this.patient, loadingOptions),
-      ]);
-
-      this.partialDataLoad = new PartialDataLoad(
-        { start: this.dataRange[0].valueOf(), end: this.dataRange[1].valueOf() },
-        { start: initialLoadingDates[0].valueOf(), end: initialLoadingDates[1].valueOf() }
-      );
-
-      combinedData = combinedData.concat(patientData, messagesNotes);
-    }
-
-    return combinedData;
   }
 
   /**
@@ -204,34 +126,26 @@ class ApiUtils {
    * @returns {Promise<PatientData>} The patient data
    */
   fetchDataRange(dateRange) {
-    const rangeToLoad = this.partialDataLoad.getRangeToLoad(dateRange);
-    if (rangeToLoad === null) {
-      this.log.warn("Empty range, we should not have ended in this situation");
-      return [];
+    const rangesToLoad = this.partialDataLoad.getMissingRanges(dateRange);
+    this.log.info("partialDataLoad updated", this.partialDataLoad.toDebug());
+
+    const promises = [];
+    for (const rangeToLoad of rangesToLoad) {
+      /** @type {GetPatientDataOptions|GetPatientDataOptionsV0} */
+      const loadingOptions = {
+        startDate: rangeToLoad.start.toISOString(),
+        endDate: rangeToLoad.end.toISOString(),
+      };
+      promises.push(this.fetchDataRangeV1(loadingOptions));
     }
 
-    /** @type {GetPatientDataOptions|GetPatientDataOptionsV0} */
-    const loadingOptions = {
-      startDate: moment.utc(rangeToLoad.start).toISOString(),
-      endDate: moment.utc(rangeToLoad.end).toISOString(),
-    };
-
-    this.log.info("Fetching data using range", {
-      request: {
-        dateRange,
-        start: new Date(dateRange.start).toISOString(),
-        end: new Date(dateRange.end).toISOString()
-      },
-      loading: {
-        dateRange: rangeToLoad,
-        start: loadingOptions.startDate,
-        end: loadingOptions.endDate,
-      },
+    return Promise.all(promises).then((values) => {
+      let r = [];
+      for (const value of values) {
+        r = r.concat(value);
+      }
+      return r;
     });
-
-    this.partialDataLoad.setRangeLoaded(rangeToLoad);
-
-    return this.haveAPIv1 ? this.fetchDataRangeV1(loadingOptions) : this.fetchDataRangeV0(loadingOptions);
   }
 
   /**
@@ -243,20 +157,6 @@ class ApiUtils {
     /** @type {[PatientData, PatientData]} */
     const [patientData, messagesNotes] = await Promise.all([
       this.api.getPatientData(this.patient, loadingOptions),
-      this.api.getMessages(this.patient, loadingOptions),
-    ]);
-    return patientData.concat(messagesNotes);
-  }
-
-  /**
-   * @param {GetPatientDataOptionsV0} loadingOptions ISO dates range
-   * @returns {Promise<PatientData>} The patient data
-   * @private
-   */
-  async fetchDataRangeV0(loadingOptions) {
-    /** @type {[PatientData, PatientData]} */
-    const [patientData, messagesNotes] = await Promise.all([
-      this.api.getPatientDataV0(this.patient, loadingOptions),
       this.api.getMessages(this.patient, loadingOptions),
     ]);
     return patientData.concat(messagesNotes);