+++ packages/blip/test/import-tests.js
@@ -12,3 +12,4 @@ import "./unit/components/chart/footer.test.js";
 import "./unit/components/messages/messages.test.js";
 import "./unit/components/messages/message.test.js";
 import "./unit/components/messages/messageform.test.js";
+import "./unit/lib/partial-data-load.test";